Word accepted bug
Solved a bug in which if you would test multiple words, some that should not be accepted are accepted.
The nextStates set retained info about the next possible states the NFA could go in and at the begging it should always be set to q0. But when you test the second word the set is already populated with other states so you do not begin from q0.
Now the set is cleared before trying to accept a word.